Fixed failure to click again after custom model loading.
The reason for this problem is that when the name parameter of the custom model configuration file is the same as the name parameter of the built-in model configuration file, the operation in the init_model_data() function will cause the loaded model information to be overwritten by the built-in model data, causing the custom model to be clicked again. Error in loading custom model: Invalid path. Error message.
The fix is also simple. User-defined models are marked with the _custom_ flag inside the application to distinguish them from built-in models.